requirements.txt
jsonpickle
geojson
shapely
pyunpack
GeoAlchemy>=0.6
requests==1.1.0
zope.interface==4.0.1
zope.schema==4.4
zope.component==4.1
zope.dottedname
z3c.schema
inflection
unidecode
pytz
isodate
dictdiffer
pytidylib
